AI Jams Update
So i tried using the getLevel method as a work around for the pitch detection not working, but it still doesn't work. I think since i am using speech recognizer, my program is not listening to any new audio in. So instead of spending this time frustrated at an issue that isnt working, I added a health function to the robot and a defeated state