Final Project: Update
I added way more features in the poseNet game. I made it so the image of the robot would fly around, I added a health bar that tracks the amount of health that is taken away when the user touches the robot, I added a countdown timer in the top right corner, and a bunch more things. I originally wanted to make it so the user had to punch and kick the robot, but unfortunately i couldnt get the cooldown time between each hit to work. Instead, I decided to make it simply touching the robot to reduce the health.